Added in a global to track the previous scene, so that when loading a specific location the character is spawned in the correct location depending on where it came from.
Also created a downstairs to the player home to test the transition between scenes. Currently working on the smoothing the transition. Currently have a placeholder code that prevents infinite loop between scenes, however need to change this as it requires the player to manually step on and off an exit to turn the exits on. Will need an "entrance" area in the players bed.